wado-compiler: extract export_adapter submodule from cm_binding
Move the export-side adapter synthesizers (synthesize_void/general/
result/async/void_stub_export_binding), the flat-args lowering
helpers (synthesize_lower_to_flat / synthesize_variant_lower_to_flat
and FlatLocal), the export-param lifter, and the
export_binding_func_name name builder into
`synthesis/cm_binding/export_adapter.rs`. task_return.rs now imports
the lowering helpers from the new submodule path.

cm_binding.rs is now down from ~10,000 to ~1,100 lines and only
holds the generate_adapters driver, the effect-owner index helpers,
and the test module.
